Moved OpenCL classes to Compute.OpenCL namespace
