+++ OpenTK.sln
@@ -78,6 +78,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OpenToolkit.Windowing.Graph
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OpenToolkit.Windowing.Desktop", "src\OpenToolkit.Windowing.Desktop\OpenToolkit.Windowing.Desktop.csproj", "{172C1E62-DA67-4203-BBCB-9DA4D9111EF0}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OpenToolkit.Compute", "src\OpenToolkit.Compute\OpenToolkit.Compute.csproj", "{7E163837-DD68-40D0-8B7B-758F7B230D5D}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -156,6 +158,10 @@ Global
 		{5F47DB55-4BFC-4832-A68F-AB36CE7431CE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5F47DB55-4BFC-4832-A68F-AB36CE7431CE}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5F47DB55-4BFC-4832-A68F-AB36CE7431CE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7E163837-DD68-40D0-8B7B-758F7B230D5D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7E163837-DD68-40D0-8B7B-758F7B230D5D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7E163837-DD68-40D0-8B7B-758F7B230D5D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7E163837-DD68-40D0-8B7B-758F7B230D5D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
