+++ src/OpenToolkit.Compute/OpenCL/ClVersionAttribute.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace OpenToolkit.Compute
+namespace OpenToolkit.Compute.OpenCL
 {
     public class ClVersionAttribute : Attribute
     {
