+++ src/OpenToolkit.Compute/OpenCL/Structs.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace OpenToolkit.Compute
+namespace OpenToolkit.Compute.OpenCL
 {
     public struct ImageFormat
     {
